writer: remove _writer suffixes in files.
This change removes the `_writer` suffix in file to shorten and simply
the file layout in order to prepare for migrating over each writers
configuration, potentially into files suffixes `_config`, for example:

```
service.go
service_config.go
service_test.go
// etc...
```
