+++ writer/multi.go
