+++ writer/multi_test.go
