+++ writer/stats.go
