+++ writer/stats_test.go
