+++ writer/trace.go
